LayoutTests/editing/inserting/insert-div-021.html

<html> 
<head>

<style>
body {
    font-size: 24px; 
}
.editing { 
    border: 2px solid red; 
    padding: 12px; 
}
div {
    border: 2px solid blue; 
    padding: 12px; 
}

</style>
<script src=../editing.js language="JavaScript" type="text/JavaScript" ></script>

<script>

function editingTest() {
    for (i = 0; i < 4; i++)
        moveSelectionForwardByCharacterCommand();
    deleteCommand();
    insertParagraphCommand();
    insertParagraphCommand();
}

</script>

<title>Editing Test</title> 
</head> 
<body contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">

Test inserting paragraphs: should see "foo", then an empty blue box, then a blue box with with "bar" in it. Fix for this bug:
<a href="rdar://problem/3924579">&lt;rdar://problem/3924579&gt;</a> REGRESSION (Mail): After deleting, hitting return key does not insert visible newline


<div style="border: none; height: 12px"></div>

<span id="test">foo</span><br>bar

<script>
runEditingTest();
</script>

</body>
</html>
